Avoid overhead of xrt::queue when single runlist
xrt::runner supports a mix of NPU and CPU runlists depending on how
runs are declared in the recipe.  Runlists are created at the boundary
of NPU/CPU individual runs. Executing a recipe implies executing all
runlist, where a CPU runlist must wait for preceeding NPU runlist to
finish (or vice versa) before starting.  This is orchestrated by an
xrt::queue, but if the recipe contains only NPU, then there will only
be one runlist and the overhead of xrt::queue should be avoided.

This PR implements the check for single runlist.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>