Break profile_impl and runner_impl, use interface base class (#8880)
* Break profile_impl and runner_impl, use interface base class

Break inheritance of profile_impl from runner_impl, which caused
strange depdendencies.

Keep runner_impl as an interface base class, derive recipe_impl and
profile_impl. Class recipe_impl manages a `recipe` and class
profile_impl manages a `profile`, which in turn is tied to a recipe.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Fix compile error

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Fix compile error

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Avoid overhead of xrt::queue when single runlist

xrt::runner supports a mix of NPU and CPU runlists depending on how
runs are declared in the recipe.  Runlists are created at the boundary
of NPU/CPU individual runs. Executing a recipe implies executing all
runlist, where a CPU runlist must wait for preceeding NPU runlist to
finish (or vice versa) before starting.  This is orchestrated by an
xrt::queue, but if the recipe contains only NPU, then there will only
be one runlist and the overhead of xrt::queue should be avoided.

This PR implements the check for single runlist.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Update comments and tweak behavior of resources::buffer

This commit changes resources::buffer behavior to always create
an xrt::bo when the recipe specifies a size for the buffer. This
is done regardless of the type of the buffer.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

---------

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>