Support strided initialization of a buffer
Extend profile json to support strided buffer initialization.

```
      "init": {
        "stride": 1,
        "value": 239,
        "begin": 0,
        "end": 524288
      }
```
Here the buffer range specified by `begin` and `end` is written with
the byte value `239` at every byte (`stride` is 1). The value can be
any uint64_t value, but must be a decimal value in order to be valid
json.  To initalize a buffer with with 0xdeadbeef, convert to decimal
and write:

```
      "init": {
        "stride": 4,
        "value": 3735928559
      }
```

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>