Notify shim when an execbo is recycled (#8889)
* Notify shim when an execbo is recycled

Add xrt_core::buffer_handle::reset() callback so that shim
can be notified when an execbo is recycled.

Exec BOs are cached in XRT coreutil and reused when creating new
xrt::run objects.  When `xrt::run::set_arg(idx, bo)` is called, shim
is notified through `xrt::buffer_handle::bind()` that the command
(execbo) associated with the run is using the `bo`.

This commit essentially adds the reverse of `bind()`, letting shim
know that even as the execbo is not destroyed through shim (it is
recycled for later use), the BOs bound to the execbo are no longer in
use by this execbo.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Fix clang-tidy warning

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

---------

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>