Enhance profile file initialization of recipe resource buffers
Initialization of a resource buffer from a file fills all the bytes of
the buffer with data from file.  It wraps around the file if necessary
to fill the bo.

The function supports initializing the buffer between iterations,
copying from the file from an offset corresponding to where
previous iteration initialization reached.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>