AIE context timeout state capture
Implement xrt::run::aie_error exception class for reporting
context health on command error.

The exception is to be thrown by xrt::wait2() on command error
when the command is an AIE command and when the command ERT
packet is populated with context health information.

This PR does not complete the plumbing, but has the necessary pieces
for implementing consumer facing reporting of context health
information.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>