common: xrt_core::device: add context manager member
There is context manager required per xrt_core::device, we add
it as a member to the xrt_core::device class.

The context manager is only going to be used by context management
which is defined in common/api/context_mgr.cpp and it is not required
to expose at all. We define a basic class in the xrt_core::device
for the context manager and make the member in xrt_core::device as
as shared pointer to this base class. The context_mgr.cpp defines
the actual derived class, and it is the one to create create and get
the context manager for the context management.

Signed-off-by: Wendy Liang <wendy.liang@amd.com>