xbtracer: add funtions messagess protobuf schema
We are using protobuf to record the XRT functions calls sequences and
its parameters, later the replayer can read the messages from the protobuf
stream to get the call sequence and the parameters, and then it can replay
the XRT calling sequence with these information.

This patch is to define the protobuf schema for the XRT functions.

Signed-off-by: Wendy Liang <wendy.liang@amd.com>