hip: core: stream: clear m_top_event when stream completes (#9084)
If a stream is set to wait for an event, all the commands
enqueued to the stream, after setting the stream waiting on
the event will be added to the commands chain to the event
so that when an event is completed, all those dependent commands
will execute. When the stream which depends on an event completed,
we should reset the event on which the stream depends on, otherwise,
the commands enqueued to the stream later will added to the event
commands chain. Since the event has completed, it will not signal
again, and thus, those commands will not run again, even we want to
rerun the stream.

Signed-off-by: Wendy Liang <wendy.liang@amd.com>