Apply 0002-Make-git-optional-for-the-build.patch (and 0006) (#9087)
Git is made optional for upstream builds.

During the build, "git" is used to query information regarding the
branch and top commit which is then embedded in the build
artifacts.

Not all builds of xrt are from the upstream trees and embedding
arbitrary git tree information is not helpful. Conditionally drop the
requirement for "git" during the build when CMake variable
XRT_UPSTREAM is defined - thus preventing the embedding of associated
information as well.

Later patch
(0006-src-CMake-version.cmake-Consistently-use-UTC-timezon.patch)
fixes how version.h is generated with semi meaningful data when Git
metadata is not available.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>