Apply 0003-Don-t-force-install-location.patch
Honor CMAKE_INSTALL_PREFIX as the installation path.

The XRT build system makes assumption about where the generated
binaries are to be installed depending on whether the target is PCIe
or edge device.

Drop these assumptions and let the caller specify if they want things
installed anywhere other than the default.

The debian patch was not entirely correct.  One should not add the
full path of CMAKE_INSTALL_PREFIX to the `install(... DESTINATION
<path>)` function, but rather rely on DESTINATION to be relative to
CMAKE_INSTALL_PREFIX.  In particular `install(EXPORT targets
.... DESTINATON <path>)`  produces wrong output if DESTINATION is an
absolute path.

In short, keep XRT_INSTALL_DIR but set it to ".".

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>