hip: function: introduce runs cache in hip functions (#9100)
In HIP stream execution flow, every kernel launch, we will need to
get a run for the command to send to the NPU firmware to execute
the kernel. In a lot of cases, between each runs, we just want to
update parameters, in such case, we don't want to create new run
each time.
And based on the evaluation, each run creation takes ~100us.
And thus, it will significantly save kernel launch time, if we can
cache previous runs.

This patch introduce a cache to cache the old runs for the function.
When calling function, in current supported case, it is the XRT hip
stream command, it will call the function's get_run() to allocate a
run if no used run in the cache, otherwise, it will return a used run
to the calling function. When the XRT hip stream command completes, that
is when it is destroyed, in the kernel_start event destructor, it will
call function's release_run() to return the run to the cache.

We don't set upper limit to the cache. In HIP stream execution flow,
stream will clears all its events when current stream complemetes, that
is whenever stream schronization is called, the commands in the stream
will be destroyed. The runs used in the streams will be returned to the
pool.

Signed-off-by: Wendy Liang <wendy.liang@amd.com>