hip: add hipSetDevice to set thread default device (#9103)
Implement HIP API to set default device of a thread.
default device is local to a thread. Currently, the default
device local to a thread is max uint32_t. It is initialized
to 0 to the main thread in global initialization. For local
thread, we will need to change to default value to 0 instetad
of max uint32_t or let the user to call hipSetDevice() to change
it, this patch enables user to call hipSetDevice() to change default
device.

Signed-off-by: Wendy Liang <wendy.liang@amd.com>