Conditionally create symlink lib -> lib64
Create symlink for `lib -> lib64` when CMAKE_INSTALL_LIBDIR is lib64.

This is work-around for test infratructure dependencies on the install
lib/ folder.  Tests should be fixed, but right now are failing on RHEL
and CentOS Linux distributions.

The symlink is created only for internal legacy XRT builds that
install to `/opt/xilinx/xrt`, it is disabled for upstream builds or
any other install prefix.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>