Remove _static suffix from static libraries on Linux
Static libraries are named same as shared libraries.  CMake targets
are unchanged to distinguish linking with static vs shared, but refer
to the renamed static library.

Sample CMakeLists.txt change:
```
if (NOT WIN32)
  # On linux use xrt_coreutil.a
  set_target_properties(xrt_coreutil_static PROPERTIES OUTPUT_NAME "xrt_coreutil")
endif()
```
Both the build and install version of the static libraries are changed, but
internal build targets that link against static libraries continue to refer
to e.g. `xrt_coreutil_static`. Under the hood CMake is aware of the renaming.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>