Amend #9231: Add minimal version info header to help compiler cache
Refactor the autogenerated version.h into two separately configured
headers, one that doesn't use timestamp (version-slim.h) and one that
does (version.h).

The XRT exported header files depend on ABI check that uses the XRT
version only.  These headers can include version-slim.h and not be
affected by a changing timestamp everytime CMake is reconfigured.
This in turn simplifies the logic for building with ccache.

The two separate headers are generated under the build tree in a
directory with same relatively location as where the corresponding
headers will be installed.

Thank you @shramov for #9231, which seeded this PR.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>