Update profile file initialization to support writing to bo range (#9253)
* Update profile file initialization to support writing to bo range

File initialization is amended with `begin` and `end` to specify
a bo range to write file data to.

```
      "init": {
        "file": "<path or repo key>",
        "skip": bytes    // skip number of bytes in file
        "begin": offset, // offset to start writing at (default: 0)
        "end": offset    // offset to end writing at (default: bo.size())
      }
```

The range defined by `[begin, end[` (default: `[0, bo.size()[`) are
the bytes of the buffer that will be populated with data from the file.

All the bytes of the buffer range are initialized regardless of the size
of the `file`.

If the file (minus skip bytes) is smaller than the buffer range, then
the file wraps around and continues to initialize the buffer range.

If the file is larger than the buffer range then only buffer range
bytes of the file are used.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Fix build error

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

---------

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>