Redo script execution in xrt-runner.exe
To reduce resources do not keep jobs (xrt::runner objects) alive after
execution.

Jobs are popped off the queue by value, thus transferring lifetime
to thread worker.  A job object is destroyed immediately after a
worker completes the job.

The script job queue is disabled until it is explicitly enabled.  This
allows for all jobs to be initialized before any one job is returned
to a worker.  If enabled up-front with new `[--asap]` switch, jobs are
returned as soon as a worker requests a job and one is available.

```
% xrt-runner --help
usage: xrt-runner.exe [options]
 [--recipe <recipe.json>] recipe file to run
 [--profile <profile.json>] execution profile
 [--iterations <number>] override all profile iterations
 [--script <script>] runner script, enables multi-threaded execution
 [--threads <number>] number of threads to use when running script (default: #jobs)
 [--dir <path>] directory containing artifacts (default: current dir)
 [--progress] show progress
 [--asap] process jobs immediately (default: wait for all jobs to initialize)
 [--report [<file>]] output runner metrics to <file> or use stdout for no <file> or '-'

% xrt-runner.exe --recipe recipe.json --profile profile.json [--iterations <num>] [--dir <path>]
% xrt-runner.exe --script runner.json [--threads <num>] [--iterations <num>] [--dir <path>]

Note, [--threads <number>] overrides the default number, where default is the number of
jobs in the runner script.

Note, [--iterations <num>] overrides iterations in profile.json, but not in runner script.
If the runner script specifies iterations for a recipe/profile pair, then this value is
sticky for that recipe/profile pair.
```

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>