Update recipe.json schema with new kernel attribute (#9281)
```
  "resources": {
    "kernels": [
      {
        ...
        "numargs": number
      }
    ]
  },
```
The `numargs` is unused by the xrt::runner class, but is added to
support using the recipe to create application code invoking a kernel
with free formed args. For example, `hipLaunchKernel` takes a void**
array of args, where the length of the array is supposed to exactly
match the number of arguments to the kernel.  The recipe `runs`
section itself may not specify all arguments, but a `hip` application
created from the recipe.json must create an array of all expected
arguments, where missing args will be populated with a default value.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>