xbmgmt2/firmware_image: Use std::vector to read std::ifstream.
This fixes the build when using `-Werror` due to `-Wmismatched-new-delete`
being triggered. Produced on at least GCC 15.2.1.

In order to avoid clang-tidy's `cppcoreguidelines-avoid-c-arrays`, use
`std::vector` to read the `std::ifstream`. This is possible since we stay in
scope during it's access, so a `std::shared_pointer` was not actually
needed.

This is already done with `std::vector<char> top` and
`std::vector<char> full` in other places of the code.

In additon some minor improvements:
* Move dtbbuf to local scope, it is never used as a member.
* Remove redundant std::move, a new std::string is constructed here.

The core of the warning was a mismatch in allocate shared_ptr arrays
without using the [] notation:

When allocating a shared_ptr and not using an array type, the delete
destructor without square brackets will be used. This leads to a
mismatch of new[] constructors that were not using delete[] destructors.

In order to properly use delete[] instead of delete destructors, the
shared_ptr template type needs to contain [] as well.

See (3) from https://en.cppreference.com/w/cpp/memory/shared_ptr/shared_ptr.html

A mismatch of these leads to undefined behavior.

https://stackoverflow.com/questions/9238731/how-serious-is-the-new-delete-operator-mismatch-error
Signed-off-by: Lubosz Sarnecki <lubosz@gmail.com>