Amend runner profile validation to support bo range validation
Add `begin` and `end` range for validation.

```
    {
      "validate": {
        "file": "ofm.bin"
        "skip": bytes    // skip number of bytes in file
        "begin": offset, // bo offset to start validation at
        "end": offset,   // bo offset to end validation at
      }
    }
```

Validating can be against the content of a file as shown in above
example.  The optional `skip` element allows skipping first bytes
of the file prior to validating against buffer data.  The range
defined by `[begin, end[` (default: `[0, bo.size()[`) are
the bytes of the buffer that will be validated against the file
data.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>