build.sh: Don't use -Werror as default.
While being a good practice to use during development, enabling -Werror
by default for deployment and user builds is a guarantee for failed
builds.

It is not feasible for development to keep up with all warnings in all
compiler versions, especially ones from the future.

To avoid the build being broken, this patch sets -Werror being disabled
as default behavior. This mimics the behavior of CMake being invoked
directly without the build.sh script.

While a zero warning policy is good, and -Werror should also be enabled
by default on the CI, it is a bad practice for deployment.

https://embeddedartistry.com/blog/2017/05/22/werror-is-not-your-friend/