hip: events: do not record the event again if event has been recorded
Before this patch, event to record will be overwritten if hipEventRecord()
is called again on the same event. However, this impelmentatino has a few issues:
* it didn't erase the event from the stream which the event has been recorded before,
  instead, it tries to erase the event from the stream it is trying to record.
* there is no lock to protect the event recording change.

This the operation to change the event recording will impact two objects, the event
itself and the stream. To simplify, this patch disable changing event recording if
event is being recorded by a stream, and add locking to protect the event state change.

Signed-off-by: Wendy Liang <wendy.liang@amd.com>