hip: event: when event is complete, commands on denpendant stream will start
This patch update the hipStreamWaitEvent() implementation to chain the dummy
event created for the stream to the actual event will be populated from other stream.
Besides the actual event will be added as the dummy event's dependency, if there is
wait event set before on the same stream, the previous wait event will also be
added as dependency of the dummy event.

When the actual event completes, it will try to submit the dummy event, the dummy
event will only starts its chained commands if all of its dependencies got resolved.

Signed-off-by: Wendy Liang <wendy.liang@amd.com>