Exclude xdp from Alveo Debian packaging
For Debian upstreaming of XRT, both NPU and ALVEO are built together
and must be able to co-exist at install time.  Since profile and debug
are conditionally compiled for ALVEO compared to NPU, they are not
included at all for upstreaming of XRT when building the ALVEO code
path.

Version MPD libraries for better package management.   This should not
affect anything as the .so remains but are now a softlink.  The deployment
package for MPD plugins have all components of the libraries.

Change components.cmake to create separate base package as part of
XRT_ALVEO.  Previously alveo component had all content.   The alveo
component package adds a dependency on the base package.  The base
package is shared with NPU and both can be installed side by side.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>