hip: implement memory prefetch async as synchronous instead
Currently, hipMemoryPrefetchAsync() implement the fetch in an async fashion,
Eventhough, it acts as synchronous most time. But caller can use it asynchronously
with event, and the memory prefetch which is doing a memory sync can wait until
dependent event is complete. However, this added overhead to create commands and push
it to the command queue of a stream which is completed for simple cases.
This patch change it to always do memory prefech synchronously instead, that is
as long as it is called, it do memory sync all the time.

Signed-off-by: Wendy Liang <wendy.liang@amd.com>