Limit maximum number of jobs in xrt-runner queue (#9342)
* Limit maximum number of jobs in xrt-runner queue

Scripted execution can specify any number of jobs, which can lead to
out-of-memory when creating and adding the xrt::runner jobs to the
queue.

This PR adds a new switch `--max-queue-size <num>` which causes
`job_queue::add(...)` to wait if current number of jobs in the
queue exceeds the specified max.  The option implies `--asap`
so that jobs are processed as soon as possible rather than all
created up-front.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Extend profile execution::mode with validate

This allows easy filtering of profile executions to any one
of the supported modes (validate, latency, throughput).

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

---------

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>