HIP events synchronization implementation update (#9328)
* hip: events: do not record the event again if event has been recorded

Before this patch, event to record will be overwritten if hipEventRecord()
is called again on the same event. However, this impelmentatino has a few issues:
* it didn't erase the event from the stream which the event has been recorded before,
  instead, it tries to erase the event from the stream it is trying to record.
* there is no lock to protect the event recording change.

This the operation to change the event recording will impact two objects, the event
itself and the stream. To simplify, this patch disable changing event recording if
event is being recorded by a stream, and add locking to protect the event state change.

Signed-off-by: Wendy Liang <wendy.liang@amd.com>

* hip: event: when event is complete, commands on denpendant stream will start

This patch update the hipStreamWaitEvent() implementation to chain the dummy
event created for the stream to the actual event will be populated from other stream.
Besides the actual event will be added as the dummy event's dependency, if there is
wait event set before on the same stream, the previous wait event will also be
added as dependency of the dummy event.

When the actual event completes, it will try to submit the dummy event, the dummy
event will only starts its chained commands if all of its dependencies got resolved.

Signed-off-by: Wendy Liang <wendy.liang@amd.com>

---------

Signed-off-by: Wendy Liang <wendy.liang@amd.com>