Fix error handling and simplify switches (#9348)
In scripted execution drop `--asap` in favor of always executing a job
as soon asa it is created.

In scriped execution drop `--max-queue-size` and dynamically adjust
the pending jobs to the number of thread workers.

Don't fail silently, ignore jobs that fail to create for any reason
and print the reason for why the jobs is ignored.

In scripted execution make sure verbosity is set to false for all
profiles, and disable printing of per job stats to console.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>