hip: device: expose hipGetDevicePropertiesR0600 API
We will need to add extern "C" to expose hipGetDevicePropertiesR0600() API,
By default ubuntu 24.04 AMD HIP bundle is of version 5.7. If we compiled
the XRT HIP on a host with 5.7 HIP installed, and then runtime, the application
runs with newer version of HIP, the compilation will fail with undefined
`hipGetDevicePropertiesR0600` as it is not exposed to user from xrt_hip library
if the library is compiled with HIP headers of a version less than 6.x.

Signed-off-by: Wendy Liang <wendy.liang@amd.com>