AIESW-16275 Support tracking of internal BO creation (#9364)
* AIESW-16275 Support tracking of internal BO creation

XRT should allow KMD to track all internal buffers created as part of
running an application.  In particular, KMD should accumulate the size
of all internal buffers.

In order to support this, XRT must channel all internal BO creation
through a single internal API that supports tracking of different use
types of buffers.  For example, if XRT creates a preemption buffer,
then the use type should reflect this, if XRT creates an instruction
buffer, then the use type should reflect this.

It is up to KMD to determine the granularity of how it collects
internal BO usage, for example, it could bucket all internal BOs into
one and report only one number, or it could be more granular and
provide a number of each use type of a buffer.

The BO flags passed from XRT to UMD (and for Windows to KMD) encodes
the use type in the flags argument.  If the use type is different from
0, it implies the buffer is an internal XRT created buffer.

This PR may require changes to SHIM, but at least on Linux the changes
have been validated through model tests, albeit, the use type is not
tracked by KMD at this point.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

* Fix comment per review

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

---------

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>