AIESW-16751 XRT umd and user mode library to support windows versioning
Add versioninfo resource to DLLs and APPs in XRT.

Configure a version-dll.rc and version-app.rc at CMake configure
time. The config file is copied from ticket where it is pre-confgured
with required fields.

The generated file is include in CMake targets.  For now only a small
subset of targets (those that are distributed) include
the versioninfo.

Subject to change.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>